Create second version of text recognition proof of concept
The VIP architecture was deemed not useful for an SDK, so only the
interactor has been left. The CiriceSDK class is acting as a
controller. The Vision wrapper is using
`withCheckedThrowingContinuation` in order to accept async/await
functionality. Request and response classes are used to avoid breaking
changes when adding new parameters to the public methods in the
controller.